+++ generate-html.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-cabal run --project-file=frontend-html.project frontend-html