+++ script/build_html.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+source ./script/utils.sh
+
+build_html