+++ script/run.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+source ./script/utils.sh
+
+version=$(get_version)
+printf "Current frontend version: %s" "$version"
+printf "\n\nRunning frontend in watch mode:\n\n"
+
+caddy run -w